+++ messages/de/voting.php
@@ -67,6 +67,7 @@ return [
     'settings_answers_present' => 'Anwesend',
     'settings_answers_presenth' => 'Für Abfragen, welche Mitglieder anwesend sind.',
     'settings_majoritytype' => 'Art der Mehrheit',
+    'settings_quorumtype' => 'Quorum',
     'settings_votepolicy' => 'Wer darf abstimmen',
     'settings_resultspublic' => 'Wer darf die Abstimmungsergebnisse in Zahlen sehen?',
     'settings_resultspublic_admins' => 'Admins',